+++ rocAL/include/augmentations/effects_augmentations/node_rain.h
@@ -22,9 +22,9 @@ THE SOFTWARE.
 
 #pragma once
 
-#include "node.h"
-#include "parameter_factory.h"
-#include "parameter_vx.h"
+#include "pipeline/node.h"
+#include "parameters/parameter_factory.h"
+#include "parameters/parameter_vx.h"
 
 class RainNode : public Node {
    public: