+++ rocAL/include/meta_data/caffe2_meta_data_reader.h
@@ -28,10 +28,10 @@ THE SOFTWARE.
 #include <memory>
 #include <variant>
 
-#include "commons.h"
-#include "meta_data.h"
-#include "meta_data_reader.h"
-#include "image_reader.h"
+#include "pipeline/commons.h"
+#include "meta_data/meta_data.h"
+#include "meta_data/meta_data_reader.h"
+#include "readers/image/image_reader.h"
 
 class Caffe2MetaDataReader : public MetaDataReader {
    public: