+++ rocAL/source/api/rocal_api_augmentation.cpp
@@ -20,11 +20,11 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 */
 
-#include "augmentations_meta_nodes.h"
-#include "augmentations_nodes.h"
-#include "commons.h"
-#include "context.h"
-#include "image_source_evaluator.h"
+#include "meta_data/augmentations_meta_nodes.h"
+#include "augmentations/augmentations_nodes.h"
+#include "pipeline/commons.h"
+#include "pipeline/context.h"
+#include "loaders/image_source_evaluator.h"
 #include "rocal_api.h"
 
 RocalTensor ROCAL_API_CALL