+++ rocAL/source/pipeline/master_graph.cpp
@@ -27,14 +27,14 @@ THE SOFTWARE.
 #include <cstring>
 #include <sched.h>
 #include <half/half.hpp>
-#include "master_graph.h"
-#include "parameter_factory.h"
-#include "ocl_setup.h"
-#include "log.h"
-#include "meta_data_reader_factory.h"
-#include "meta_data_graph_factory.h"
-#include "randombboxcrop_meta_data_reader_factory.h"
-#include "node_copy.h"
+#include "pipeline/master_graph.h"
+#include "parameters/parameter_factory.h"
+#include "device/ocl_setup.h"
+#include "pipeline/log.h"
+#include "meta_data/meta_data_reader_factory.h"
+#include "meta_data/meta_data_graph_factory.h"
+#include "meta_data/randombboxcrop_meta_data_reader_factory.h"
+#include "augmentations/node_copy.h"
 
 using half_float::half;
 
@@ -906,7 +906,7 @@ void MasterGraph::output_routine() {
                 break;
             auto full_batch_data_names = _loader_module->get_id();
             auto decode_data_info = _loader_module->get_decode_data_info();
-            auto CropImageInfo = _loader_module->get_crop_image_info();
+            auto crop_image_info = _loader_module->get_crop_image_info();
 
             if (full_batch_data_names.size() != _user_batch_size)
                 WRN("Master Graph: Names count does not equal batch_size" + TOSTR(full_batch_data_names.size()))
@@ -937,7 +937,7 @@ void MasterGraph::output_routine() {
                 output_meta_data = _augmented_meta_data->clone(!_augmentation_metanode);  // copy the data if metadata is not processed by the nodes, else create an empty instance
                 if (_meta_data_graph) {
                     if (_is_random_bbox_crop) {
-                        _meta_data_graph->update_random_bbox_meta_data(_augmented_meta_data, output_meta_data, decode_data_info, CropImageInfo);
+                        _meta_data_graph->update_random_bbox_meta_data(_augmented_meta_data, output_meta_data, decode_data_info, crop_image_info);
                     } else {
                         _meta_data_graph->update_meta_data(_augmented_meta_data, decode_data_info);
                     }