+++ rocAL/source/readers/image/cifar10_data_reader.cpp
@@ -21,12 +21,12 @@ THE SOFTWARE.
 */
 
 #include <cassert>
-#include <commons.h>
+#include "pipeline/commons.h"
 #include <cstring>
 #include <algorithm>
-#include "cifar10_data_reader.h"
-#include <file_source_reader.h>
-#include "filesystem.h"
+#include "readers/image/cifar10_data_reader.h"
+#include "readers/image/file_source_reader.h"
+#include "pipeline/filesystem.h"
 
 CIFAR10DataReader::CIFAR10DataReader() {
     _src_dir = nullptr;